Add WAMP ROSbotXL teleoperation plugin
This file starts from the teleop_twist_keyboard.py file found in the ROSbotXL Github repository. In this file the teleop operation has been tweaked in order to make the interaction possibile between S4T/LR and the robot from remote. There is also the possibility to control the robot via two analog virtual pads which behave like a real joystick. In this case too the implementation is a little modification of the already existen found in the rosbot repos.